# enroute-cp/docs/swagger.yaml
definitions:
...
  webhttp.Upstream:
    properties:
      upstream_hc_healthythresholdcount:
        type: string
      upstream_hc_host:
        type: string
      upstream_hc_intervalseconds:
        type: string
      upstream_hc_path:
        type: string
      upstream_hc_timeoutseconds:
        type: string
      upstream_hc_unhealthythresholdcount:
        type: string
      upstream_ip:
        type: string
      upstream_name:
        type: string
      upstream_port:
        type: string
      upstream_strategy:
        type: string
      upstream_validation_cacertificate:
        type: string
      upstream_validation_subjectname:
        type: string
      upstream_weight:
        type: string
    type: object
info:
  contact:
    name: API Support
    url: https://saaras.io/
  description: APIs to configure multiple envoy proxies
  license: {}
  title: Enroute API
  version: "1.0"
paths:
...
  /service/{service_name}/route/{route_name}/upstream:
    get:
      consumes:
      - application/json
      description: Get all upstreams associated with a service route
      parameters:
      - description: Name of service
        in: path
        name: service_name
        required: true
        type: string
      - description: Name of route
        in: path
        name: route_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - ApiKeyAuth: []
      summary: Get all upstreams associated with a service route
      tags:
      - service
      - route
      - upstream
  /service/{service_name}/route/{route_name}/upstream/{upstream_name}:
    get:
      consumes:
      - application/json
      description: Disassociate an upstream with a service route. This does not delete
        the upstream.
      parameters:
      - description: Name of service
        in: path
        name: service_name
        required: true
        type: string
      - description: Name of route
        in: path
        name: route_name
        required: true
        type: string
      - description: Name of upstream to disassociate with service route
        in: path
        name: upstream_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - ApiKeyAuth: []
      summary: Disassociate an upstream with a service route
      tags:
      - service
      - route
      - upstream
...
  /upstream:
    get:
      consumes:
      - application/json
      description: Get a list of all upstreams
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - ApiKeyAuth: []
      summary: List all upstreams
      tags:
      - upstream
    post:
      consumes:
      - application/json
      parameters:
      - description: Upstream to create
        in: body
        name: Upstream
        required: true
        schema:
          $ref: '#/definitions/webhttp.Upstream'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - ApiKeyAuth: []
      summary: Create an upstream
      tags:
      - upstream
  /upstream/{upstream_name}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Name of upstream to delete
        in: path
        name: upstream_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - ApiKeyAuth: []
      summary: Delete an upstream
      tags:
      - upstream
    get:
      consumes:
      - application/json
      parameters:
      - description: Name of upstream to delete
        in: path
        name: upstream_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - ApiKeyAuth: []
      summary: Get info of an upstream
      tags:
      - upstream
    patch:
      consumes:
      - application/json
      parameters:
      - description: Upstream to update
        in: body
        name: Upstream
        required: true
        schema:
          $ref: '#/definitions/webhttp.Upstream'
          type: object
      - description: Name of upstream to update
        in: path
        name: upstream_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - ApiKeyAuth: []
      summary: Update an upstream
      tags:
      - upstream
  /upstream/{upstream_name}/route:
    get:
      consumes:
      - application/json
      parameters:
      - description: Name of upstream
        in: path
        name: upstream_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - ApiKeyAuth: []
      summary: Get routes to which this upstream is associated
      tags:
      - upstream
      - route
  /upstream/copy/{upstream_name_src}/{upstream_name_dst}:
    get:
      consumes:
      - application/json
      description: Make a copy from upstream_name_src to upstream_name_dst
      parameters:
      - description: Name of upstream
        in: path
        name: upstream_name_src
        required: true
        type: string
      - description: Name of upstream
        in: path
        name: upstream_name_dst
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - ApiKeyAuth: []
      summary: Copy an upstream
      tags:
      - upstream
      - operational-verbs
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"